+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Add horizontal scrollbar in the input editor (#316)
 * Change the format of the input file, to make it more robust and adaptative (#332)
 * Input Editor window can get behind main window (#334, #49)
+* Home/End keys in the input editor move to top/bottom row (#335)
 
 ### Fixed